Merged the changes on the main CVS branch into the alias branch and
Estimated hours taken: 20

Merged the changes on the main CVS branch into the alias branch and
fixed all the conflicts.  This is to allow the alias branch to compile
after after the split up of mercury_builtin.m.

Note: this version only bootstraps with -O-1, but that's better than
not compiling at all.