Make the selection of candidates for reuse strict: don't indicate
Estimated hours taken: 0.10

Make the selection of candidates for reuse strict: don't indicate
different constructions to reuse the same dead cell (which is absurd
of course). The reason why it was differently before was to
simply report indication on where one could reuse what. But now
we have to really decide. (Later on, that decision could be made
a bit smarter: now it's just based on a LIFO strategy... )

sr_reuse.m:
	Only reuse a dead cell if it is not already reused by a
	previous construction.
	( including preparatory code for allowing other cons-id's
	than the cons being constructed to be reused for that
	construction)