A temporary fix of a problem which has a bit more of theoretical roots.
Estimated hours taken: 5

A temporary fix of a problem which has a bit more of theoretical roots.
The problem was that some conditions, which should propagate, didn't
(i.e. became `always'). The reason for this is that when translating
nodes which are being reused in the called predicate to the current
environment, it really looks at the full nodes and not at the subparts
of it. If the nodes itself are not related to a headvariable, then
the condition becomes `always'.
This must have been a bug which was always there but only showed up
now.

sr_indirect.m:
	extra verbosity.
	Bugfix: when updating a reuse_condition to the current environment,
	make sure to normalize the aliases first (e.g.: the tail of
	a list will be mapped unto list itself) so that you really
	can check for how the nodes a call is reusing are related to
	the headvariables.
	This seems to me as a temporary fix and should be once checked
	if more time is available.

sr_direct.m:
	extra verbosity.