Instead of relying on whatever unheld convention, add a new field
Estimated hours taken: 0.5

Instead of relying on whatever unheld convention, add a new field
to the proc_info which will hold the original list of headvariables.
No predicate is provided to change this original list as we don't
want them to be changeable by anything.
This information is necessary for the alias and reuse analysis
in order to output the correct sequence of headvariables in terms
of which the alias and reuse-information is expressed.

hlds_pred.m:
	Add new field, real_head_vars, to the proc_info, and
	the predicate to access this information.

pa_run.m:
structure_reuse.m:
	Use the new predicate for fetching the real headvars
	from the proc_info.