Change the deep profile analyser to use a dense bitset
Estimated hours taken: 1

Change the deep profile analyser to use a dense bitset
(using an array of ints) rather than a bbb-tree set
to store the "visited" set of nodes during the DFS
in the code that does the topological sort of the call graph.
This took the running time from 95s to 8.4s.