Kind of bugfixes.
Estimated hours taken: 6

Kind of bugfixes.
1. The fixpoint-table used for keeping track of the
alias-results used to compare two subsequent abstract substitutions
on exact equality. This was correct as long as one was working with
normal selectors. In the presence of type-selectors and type-widening
it requires a less-or-equal comparison or else the analysis might
never stop (in some run, the abstract substitutions might be widened,
whereas in the next run it will not be the case... they are not equal
as such, but one will be less-or-equal wrt the other).
2. Correct the normalization-step. Normalization was not adapted to
deal with type-selectors and hence produced wrong results.
3. cleaning up here and there... + small changes...


pa_run.m:
pa_util.m:
	Adaptation of the fixpoint table used for the alias abstract
	substitutions: instead of testing on equality, we now test
	on less-or-equal'ness.

pa_selector.m:
	Correction of the normalization-predicate: also handle type-selectors
	adequately (instead of just ignoring them).

pa_alias.m:
	Cleanup + eliminate use of pa_selector__less_or_equal/3 which
	may only be used when no type-selectors are involved.

pa_alias_as.m:
	Renaming of leq into less_or_equal.

pa_datastruct.m:
	New predicate for creating a datastruct given a program variable
	and selector.