Bugfix. When a procedure has conditional reuse, even if each of the
Estimated hours taken: 7 (6 for peter tracking the bug, 1 for fixing it)

Bugfix. When a procedure has conditional reuse, even if each of the
conditions are satisfied within the callers context, still reuse
is not allowed if the nodes which are being reused in the procedure
are aliased to each other within the callers context.

sr_data.m:
	The bug can be fixed by controlling whether none of reuse
	nodes is aliased to any of the other reuse-nodes (or itself)
	within the callers-context.