Major cleanup and associated bug fixes. The major changes are that
Estimated hours taken: 40
Branches: deep2

Major cleanup and associated bug fixes. The major changes are that
all proc_static and call_site_static structures are written out,
proc_static and call_site_static structures now contain more information,
this extra information is used to print source information in the browser,
the browser groups together all CSDs that came from the same call site,
there is a more friendly front end and with more meaningful summary options,
and the root of the call graph is now a proc_dynamic for the runtime system's
calls to the init, main and terminate routines.