Bugfix. The aliases didn't propagate correctly:
Estimated hours taken: 1 (+ 24 hours to locate and filter out the bug --peter)
Branches: reuse

Bugfix. The aliases didn't propagate correctly:
		% deconstruct
		List0 => [ FirstElement | Rest ]
If the elements of List0 are already aliased to something else, then
the elements of Rest might also be aliased to that something else. This
wasn't the case, and resulted in wrong reuses (things were reused while
being needed).
The reason of this is that while extending aliases with new sets of aliases,
recursive types were not correctly propagated. The solution to this consists
of a correction of the notion of less-then-or-equal for selectors, in such
a way that the selectors are first normalized, and only then compared:
	S1 <= S2 <=> NormalisedS1 <= NormalizedS2

pa_selector.m:
	First normalize the selectors before comparing them.