sr_choice used to allow the following reuse situation:
Estimated hours taken: 0.3
Branches: reuse

sr_choice used to allow the following reuse situation:

:- type t ---> f(int, int); g(int, int).
:- pred transform(t::in, t::out) is det.
transform(A, B):-
	(
		not( A = f(_, _) )	% A considered dead after this goal
	->
		B = g(1, 2)		% reuse A
	;
		B = g(3, 4)
	).

A might die in the negated goal, but when leaving the negation, A is
still dead, but you cannot always be sure of the functor A is having. So
reuse of A shouldn't be allowed.

sr_choice.m:
	Don't take into account cells dying in negated goals.