Bugfix. If a procedure has conditional reuse, it is split up into 2
Estimated hours taken: 1

Bugfix. If a procedure has conditional reuse, it is split up into 2
versions: a version having all reuse, and a version having all
unconditional reuse (i.e. reuse which is always safe to do,
independent of the caller). Even if a reuse has no conditional reuse, the
version is still processed to allow all unconditional reuse.
When introducing the concept of "potential_reuse" as an extra step
before the actual reuses, things went wrong, and the unconditional
reuses in the versions with pure unconditional reuses, were not
marked as actual reuses, hence missing a huge part of all the
reuses.
Solution: when a reuse is known to be unconditional, don't go mark
it as "potential reuse", but directly as a plain reuse tout-court.

sr_choice_graphing.m:
sr_indirect.m:
	When reuse is detected, check whether it is conditional
	or not. If yes: produce a potential_reuse(...), or else,
	simply move on to reuse(...).