Move the lifo selection criterium to be part of the sr_choice_graphing
Estimated hours taken: 2
Branches: reuse

Move the lifo selection criterium to be part of the sr_choice_graphing
implementation. The only difference between a lifo approach and a full graph
approach now is that if there are multiple possible reuses in each child of a
conjunction, then you need to pick the first one, instead of the one with the
largest value.
This change makes the lifo to be a hybrid form between what it used to be, and
the full graph-based approach. The main difference is that after
matching-construction-deconstructions have been identified, we still first
select the one with the best value, instead of picking the first match that was
found. As this can only improve the lifo-heuristic, it is kept this way.

The next step consists in removing the "random"-selection strategy, and with
it, module "sr_choice".

compiler/sr_choice.m:
compiler/sr_choice_graphing.m:
compiler/sr_direct.m: