Bugfix!
Estimated hours taken: 10
Branches: reuse

Bugfix!
According to the theory (phd Nancy), reuse condition information consists of
three types of information:
	1. the datastructures approximating the memory cells that can
	potentially be reused;
	2. the datastructures that are definitely live at the reuse site;
	3. the set of aliases to take into account at that reuse site.

We always approximated (2) by a set of variables, yet the implVementation of
that approximation was wrong. Indeed, it is not enough to take all the local in
use information and project it on the headvars, it first has to be extended
w.r.t. the present aliasing information (cf theory phd Nancy). By using the
pure variable representation, the result of extending and then projecting
becomes a huge overestimation of what is actually in use at that reuse site,
hence resulting in stricter reuse conditions and thus an overall worse memory
reuse behavior.  It therefore becomes mandatory to switch to a
datastructure-representation of the inuse-part of the reuse information.

This diff is therefore a bugfix (the in use information is now extended w.r.t.
aliasing before being projected) combined with an increase in the analysis
precision wrt that bugfix by conforming the analysis implementation to the
theory (phd Nancy): the in use part of the structure reuse is now represented
using data structures instead of pure variables.

As a consequence of this bugfix, the results obtained with the icfp-benchmark
are slightly worse (about 1-2% more memory cells are needed to run the
optimised code). Indeed, the bugfix adds more information to be taken
into account when verifying the safeness of some of the reuses, which means
that some reuses are now identified as unsafe instead of safe, hence less
overall reuses.

sr_data.m:
	When computing the reuse information generated at a reuse site,
	take into account the local in use information represented as data
	structures instead of pure variables, and extend it using the local
	aliases before projecting it onto the headvariables. This conforms to
	the theory.

pa_alias.m:
pa_alias_as.m:
prog_io_pasr.m:
sr_dead.m:
sr_indirect.m:
sr_live.m:
	All InUse information is now represented using data structures during
	the analysis.
	Changes supporting the main change in sr_data.